+++ src/rent/complete.js
@@ -913,7 +913,11 @@ require(['jquery',
       }
 
       // Configure Telephone with prefix
-      var countryCode = commonUI.intlTelInputCountryCode();
+      //var countryCode = commonUI.intlTelInputCountryCode();
+      var countryCode = model.configuration.countryCode;
+      if (typeof countryCode === 'undefined' || countryCode == null) {
+        countryCode = commonUI.intlTelInputCountryCode(); 
+      }
 
       if ($('#customer_phone').length) {
         $("#customer_phone").intlTelInput({